CR-2766 - AME-2631 - Monitor when the reaper runs and the number of expired tokenspwd

+++ openam-core/src/main/java/org/forgerock/openam/monitoring/cts/CtsCRUDOperationsEntryImpl.java
@@ -121,7 +121,7 @@ public class CtsCRUDOperationsEntryImpl extends CtsCRUDOperationsEntry {
             throw new InvalidSNMPQueryException();
         }
 
-        return ((Double) monitoringStore.getAverageOperationsPerPeriod(null, getCTSOperation())).longValue();
+        return (long) monitoringStore.getAverageOperationsPerPeriod(null, getCTSOperation());
     }
 
     /**
